moved test directories
