+++ __tests__/components/Splash.test.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
-import Splash from '../src/components/Splash'
+import Splash from '../../src/components/Splash'
 
 describe("<Splash />", () => {
   it('renders correctly', () => {
